modifing project structure
