+++ src/app/app.component.css
@@ -1,4 +1,4 @@
 .home_background{
-    background-image: url(../assets/images/bg.png);
-    background-repeat: no-repeat;
+    /* background-image: url(../assets/images/bg.png);
+    background-repeat: no-repeat; */
   }
