+++ src/app/app.component.html
@@ -1,3 +1,3 @@
-<!-- <app-header></app-header> -->
+<app-header *ngIf="currentUrl != '/' "></app-header>
 <router-outlet></router-outlet>
-<!-- <app-footer></app-footer> -->
+<app-footer *ngIf="currentUrl != '/' "></app-footer>
