+++ src/app/app.component.ts
@@ -1,7 +1,5 @@
 import { Component, OnInit } from '@angular/core';
-import * as $ from 'jquery';
-import { FormBuilder, FormGroup, Validators} from '@angular/forms';
-import { Router, Route, ActivatedRoute } from '@angular/router';
+import { Router, ActivatedRoute, NavigationEnd, Event } from '@angular/router';
 import { Ng4LoadingSpinnerService } from 'ng4-loading-spinner';
 import { HeaderService } from './header.service';
 
@@ -13,16 +11,24 @@ declare var $: any;
 })
 export class AppComponent{
   show=true;
+  currentUrl: string;
   constructor(private activeRoute:ActivatedRoute,
-    private spinnerService: Ng4LoadingSpinnerService,private headerService: HeaderService) {
+    private spinnerService: Ng4LoadingSpinnerService, 
+    private headerService: HeaderService, 
+    public router: Router) {
     this.spinnerService.show();
     setTimeout(() => this.spinnerService.hide(),800);
-   
    }
   title = 'hemkart';
+
   ngOnInit() {
     this.headerService.title.subscribe(title => {
       this.title = title;
+    });
+    this.router.events.subscribe((event: Event) => {
+      if (event instanceof NavigationEnd ) {
+        this.currentUrl = event.url;
+      }
     });
   }
 }
