+++ src/app/app.module.ts
@@ -12,14 +12,16 @@ import { LandingComponent } from './layout/landing/landing.component';
 import { ProductModule } from './product/product.module';
 import { HttpClientModule } from '@angular/common/http';
 import { HeaderService } from './header.service';
-import { SharedModule } from './shared';
+import { LoginComponent } from './layout/login/login.component';
+
 @NgModule({
   declarations: [
     AppComponent,
     HomeComponent,
     HeaderComponent,
     FooterComponent,
-    LandingComponent
+    LandingComponent,
+    LoginComponent
   ],
   imports: [
     BrowserModule,
@@ -30,7 +32,6 @@ import { SharedModule } from './shared';
     Ng4LoadingSpinnerModule.forRoot(),
     AppRoutingModule,
     ProductModule,
-    SharedModule  
   ],
   providers: [HeaderService],
   bootstrap: [AppComponent]
