+++ src/app/layout/home/home.component.css
@@ -120,7 +120,7 @@
   display: none;
 }
  .container_new_hole_div {
-  margin-top: 66px;
+  /* margin-top: 66px; */
 }
 .navbar-fixed-top.scrolled1 {
 background-color: #fff !important;
