+++ src/app/layout/login/login.component.html
@@ -1,0 +1,3 @@
+<p>
+  login works!
+</p>
