--- src/app/shared/index.ts
@@ -1,3 +1,0 @@
-export * from './layout';
-export * from './shared.module';
-
