--- src/app/shared/layout/footer.component.css
@@ -1,120 +1,0 @@
-/*footer css starting*/
-.fooder_bg {
-    background-color: #f2f2f2;
-    /* box-shadow: 0px -4px 3px rgba(156, 152, 152, 0.75); */
-    /* background-color: #ffffff; */
-    /* background-image: linear-gradient(180deg, #CCC 2%, #ffffff 95%); */
-    /* background: url(../images/shadow.png) repeat-x; */
-}
-.footer_hole_div{
-	border-bottom: 1px solid gray;
-}
-.margin_top_bottom_class {
-    padding: 39px 0px;
-    /* border-bottom: 1px solid #fff; */
-}
-.footet_head_h1 {
-    font-family: KlarnaHeadline,"Helvetica Neue",Arial,Helvetica,sans-serif;
-    color: #000;
-    font-size: 15px;
-    text-transform: uppercase;
-    font-weight: 600;
-}
-.liclass {
-    font-family: KlarnaSans,"Helvetica Neue",Arial,Helvetica,sans-serif;
-    list-style: none;
-    color: #14130f;
-    font-weight: 500;
-    font-size: 15px;
-}
-.link_a_class {
-    list-style: none;
-    color: #14130f;
-    font-weight: 500;
-    font-size: 15px;
-}
-.link_a_class:hover {
-    list-style: none;
-    color: #888784;
-    font-weight: 500;
-    font-size: 15px;
-    text-decoration: none;
-}
-/*.link_a_class:hover {
-    list-style: none;
-    color: #888784;
-    font-weight: 500;
-    font-size: 15px;
-    text-decoration: none;
-}*/
-.link_a_class:hover, .link_a_class:focus, .link_a_class:active {
-    color: #fff;
-    text-decoration: none !important;
-}
-.icon_span {
-    font-size: 21px;
-    padding: 4px 10px;
-    color: #808080;
-    border: 1px solid #808080;
-    border-radius: 63px;
-}
-.icon_span:hover {
-    border: 1px solid #790303;
-    background-color: #790303;
-    color: #fff;     
-    text-decoration: none;
-}
-.img-apple {
-    margin-bottom: 10px;
-    border: 1px solid #fff;
-    border-radius: 11px;
-}
-.img_iphone {
-    border: 1px solid #fff;
-    border-radius: 11px;
-}
-.footer_img_div {
-    padding: 23px 0px;
-}
-.footer_logo{
-	height: 33px;
-}
-.copy_p {
-    font-family: KlarnaHeadline,"Helvetica Neue",Arial,Helvetica,sans-serif;
-    font-size: 15px;
-    letter-spacing: 0.1px;
-    color: #808080;
-    padding: 23px 0px;
-    margin: 0px !important;
-    font-weight: 500;
-    text-align: right;
-}
-.com_link {
-    font-size: 15px;
-    letter-spacing: 0.1px;
-    color: #808080;
-    padding: 23px 0px;
-    margin: 0px !important;
-    font-weight: 500;
-}
-/*footer media query*/
-@media only screen and (max-width: 767px) and (min-width: 320px){
-.margin_top_bottom_class {
-    padding: 18px 0px !important;
-    /* border-bottom: 1px solid #fff; */
-    text-align: center;
-}
-.footer_cont_padding {
-    padding: 0px !important;
-}
-.footer_img_div {
-    text-align: center;
-    padding: 12px 0px;
-}
-.copy_p {
-    font-size: 15px;
-    letter-spacing: 0.1px;
-    color: #808080;
-    padding: 0px !important;
-}
-}
