--- src/app/shared/layout/footer.component.html
@@ -1,52 +1,0 @@
-<div class="container-fluid fooder_bg">
-  <div class="container">
-    <div class="contact_div">
-      <div class="col-md-12 col-sm-12" style="border-bottom: 1px solid gray;">
-        <div class="col-md-3 col-sm-6 margin_top_bottom_class">
-          <div class="footer_cont_head">
-            <h1 class="footet_head_h1">FÖRETAG</h1>
-            <li class="liclass">Onvision AB</li>
-            <li class="liclass"><a class="link_a_class" href="tel:559004-8269">559004-8269</a></li>
-            <li class="liclass"><a class="link_a_class" href="#">support@xn--hemkrtochklart-ypb.se</a></li>
-          </div>
-        </div>
-        <div class="col-md-3 col-sm-6 margin_top_bottom_class">
-          <div class="footer_cont_head">
-            <h1 class="footet_head_h1">KONTAKTA</h1>
-            <li class="liclass"><a class="link_a_class" href="#"> Villkor</a> </li>
-            <li class="liclass"><a class="link_a_class" href="#"> Vanliga frågor</a></li>
-            <li class="liclass"><a class="link_a_class" href="#"> Restauranger (Västerås)</a></li>
-            <li class="liclass"><a class="link_a_class" href="#"> Restauranger (Eskilstuna)</a></li>
-          </div>
-        </div>
-        <div class="col-md-3 col-sm-6 margin_top_bottom_class">
-          <div class="footer_cont_head">
-            <h1 class="footet_head_h1">SOCIAL</h1>
-            <a class="social_a" target="_blank" href="https://www.facebook.com/pages/Hemk%C3%B6rtochklartse/1113588005318795?fref=ts" style="    padding-right: 3px;">
-              <span class="icon_span"><i class="fa fa-facebook-f"></i></span>
-            </a>
-            <a class="social_a" target="_blank" href="https://www.instagram.com/">
-              <span class="icon_span"><i class="fa fa-instagram"></i></span>
-            </a>
-          </div>
-        </div>
-        <div class="col-md-3 col-sm-6 margin_top_bottom_class">
-          <div class="footer_cont_padding">
-            <h1 class="footet_head_h1">Hemkört&Klart i mobilen</h1>
-            <a href="#"><img class="img-apple " src="./assets/images/app_store_btn.png"></a><br>
-            <a href="#"><img class="img_iphone" src="./assets/images/google_play_btn.png"></a>
-          </div>
-        </div>
-      </div>
-    </div>
-    <div class="footer_div">
-      <div class="col-md-4 col-sm-4 footer_img_div">
-        <a href=""><img style="height: 33px;" src="./assets/images/1428568091-Attachment-1.png"></a>
-      </div>
-      <div class="col-md-8 col-sm-8 copy_write_div">
-        <p class="copy_p"> upphovsrätt &copy; 2018 Hemkört&Klart | Utvecklad av <a class="com_link" target="_blank"
-            href="http://inspirepro.eu/">Inspirepro Product Technology AB</a></p>
-      </div>
-    </div>
-  </div>
-</div>
