--- src/app/shared/layout/index.ts
@@ -1,2 +1,0 @@
-export * from './footer.component';
-export * from './header.component';
